refactor(api): move competence-evaluation-controller
