+++ api/lib/routes.js
@@ -22,7 +22,6 @@ import * as certificationReports from './application/certification-reports/index
 import * as certificationIssueReports from './application/certification-issue-reports/index.js';
 import * as certifications from './application/certifications/index.js';
 import * as challenges from './application/challenges/index.js';
-import * as competenceEvaluations from './application/competence-evaluations/index.js';
 import * as complementaryCertifications from './application/complementary-certifications/index.js';
 import * as complementaryCertificationCourseResults from './application/complementary-certification-course-results/index.js';
 import * as countries from './application/countries/index.js';
@@ -85,7 +84,6 @@ const routes = [
   certificationIssueReports,
   certifications,
   challenges,
-  competenceEvaluations,
   complementaryCertifications,
   complementaryCertificationCourseResults,
   countries,
