+++ api/src/evaluation/application/competence-evaluations/competence-evaluation-controller.js
@@ -1,6 +1,6 @@
-import { usecases } from '../../domain/usecases/index.js';
-import * as competenceEvaluationSerializer from '../../infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
-import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
+import { usecases } from '../../../../lib/domain/usecases/index.js';
+import * as competenceEvaluationSerializer from '../../../../lib/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 
 const startOrResume = async function (request, h, dependencies = { competenceEvaluationSerializer }) {
   const userId = request.auth.credentials.userId;
