+++ api/src/evaluation/routes.js
@@ -1,0 +1,5 @@
+import * as competenceEvaluationsRoutes from './application/competence-evaluations/index.js';
+
+const evaluationRoutes = [competenceEvaluationsRoutes];
+
+export { evaluationRoutes };
